pyproject.toml
[project]
name = "azla-mcts"
version = "0.1.0"
description = "AlphaZero-Latent Agreement MCTS with TRL PPO and LoRA for Gemma"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "AZLA Team"}]
dependencies = [
  "pydantic>=2.8",
  "pyyaml>=6.0",
  "transformers>=4.43",
  "accelerate>=0.28",
  "peft>=0.12",
  "trl>=0.8.6",
  "numpy>=1.24",
  "scipy>=1.11",
]

[project.scripts]
azla = "azla.cli:main"

[project.optional-dependencies]
quantization = [
  "bitsandbytes>=0.43",
]
dev = [
  "pytest>=7.4",
]

[tool.uv]
package = true
dev-dependencies = ["pytest>=7.4"]
scripts.main = "accelerate launch -m azla.cli --config configs/default.yaml --run train"
scripts.infer = "python scripts/compare_infer.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.black]
line-length = 100
